update dos arquivos e correção modularização
